Enable running rake target without all
Now when you run `rake pdf` it does not build all the books, but throws
some strange error. I realize that you should run `rake pdf[all]` but I
think it is implied that `rake pdf` should be called for all books.

It is a simple check if the book argument is null, then also consider it
in all target.